Much better tests for BundleDeserializer
Tests are based around test scenarios and just one test method.
Each scenario has three variables: json_text, selectors and validators.
json_text contains the whole text of the bundle to deserialize. 
selectors is a dictionary of variable name to lambda mapping. Each lambda is
called with the deserialized object and the result of the computation is stored
as attribute. Finally validators is an array of lambdas, each lambda is called
with self (the test object) and object created by running all selectors.

The design allows for tight coupling of json_text and tests (using
self.assertXXX).  Selectors allow to shorten attribute traversal as something
deeply nested can be 'selected' just once.