Add schema and object tracing to JSON validator
This patch is fairly small, specifically id does not change the API nor
invalidate any existing unit tests. The advantage is that each ValidationError
now has a trace of objects/schemas traversed in 'object_expr and 'schema_expr'
so the error reporting can be much better. The old message is now deprecated
(it was rather useless as it kept spitting out lots of JSON which made the
output unreadable) and a new_message field provides a concise overview of the
failure that is meant to be understood by mortals.