Change Schema.type to return true value
Schema.type used to return a list of types regardless of the actual
schema definition. This behavior is now removed and the true definition
is returned (after some validation).

In addition the list-of-types, aka the union type is now properly
validated. The union type is only valid when at least two possible types
are defined.

Unit tests needed some fixes, hopefully user code will need minor
adjustments after this incompatible change. The schema class is not used
widely outside of the validator.